+++ groups/bsl/bslstl/bslstl_map.t.cpp
@@ -4747,7 +4747,7 @@ void TestDriver<KEY, VALUE, COMP, ALLOC>::testCase7_1()
 
         for (int ti = 0; ti < NUM_SPECS; ++ti) {
             const char *const SPEC   = SPECS[ti];
-            const size_t      LENGTH = (int) strlen(SPEC);
+            const size_t      LENGTH = strlen(SPEC);
             TestValues VALUES(SPEC);
 
             if (verbose) {
@@ -4867,7 +4867,7 @@ void TestDriver<KEY, VALUE, COMP, ALLOC>::testCase7()
 
         for (int ti = 0; ti < NUM_SPECS; ++ti) {
             const char *const SPEC   = SPECS[ti];
-            const size_t      LENGTH = (int) strlen(SPEC);
+            const size_t      LENGTH = strlen(SPEC);
 
             if (verbose) {
                 printf("\nFor an object of length " ZU ":\n", LENGTH);
@@ -5504,7 +5504,7 @@ void TestDriver<KEY, VALUE, COMP, ALLOC>::testCase3()
             const int         LINE   = DATA[ti].d_line;
             const char *const SPEC   = DATA[ti].d_spec;
             const int         INDEX  = DATA[ti].d_index;
-            const size_t      LENGTH = (int)strlen(SPEC);
+            const size_t      LENGTH = strlen(SPEC);
 
             Obj mX(&oa);
