+++ groups/bsl/bslstl/bslstl_queue.t.cpp
@@ -1879,7 +1879,7 @@ void TestDriver<VALUE, CONTAINER>::testCase7()
 
         for (int ti = 0; ti < NUM_DATA; ++ti) {
             const char *const SPEC   = DATA[ti].d_spec;
-            const size_t      LENGTH = (int) strlen(SPEC);
+            const size_t      LENGTH = strlen(SPEC);
 
             if (verbose) {
                 printf("\nFor an object of length " ZU ":\n", LENGTH);
@@ -2419,7 +2419,7 @@ void TestDriver<VALUE, CONTAINER>::testCase3()
             const int         LINE   = DATA[ti].d_line;
             const char *const SPEC   = DATA[ti].d_spec;
             const int         EXPR   = DATA[ti].d_return;
-            const size_t      LENGTH = (int)strlen(SPEC);
+            const size_t      LENGTH = strlen(SPEC);
 
             Obj mX(&oa);
