+++ groups/bsl/bslstl/bslstl_unorderedmap.t.cpp
@@ -7452,8 +7452,8 @@ void TestDriver<KEY, VALUE, HASH, EQUAL, ALLOC>::testCase7()
         const int NUM_SPECS = sizeof SPECS / sizeof *SPECS;
 
         for (int ti = 0; ti < NUM_SPECS; ++ti) {
-            const char *const SPEC        = SPECS[ti];
-            const size_t      LENGTH      = (int) strlen(SPEC);
+            const char *const SPEC   = SPECS[ti];
+            const size_t      LENGTH = strlen(SPEC);
 
             if (verbose) {
                 printf("\nFor an object of length " ZU ":\n", LENGTH);
